pour fix des bugs

+++ Controller/ControllerUtilisateur.php
@@ -12,12 +12,20 @@ class ControllerUtilisateur {
 
 
     public static function readAll() {
-        $tab_uti = ModelUtilisateur::selectAll();     //appel au modèle pour gerer la BD
-        $controller = static::$object;
-        $view = "list";
-        $pagetitle = "Tout les utilisateurs";
-        $list = File::build_path(array("View","view.php"));
-        require $list;
+        if ($_SESSION['admin'] == 1) {
+            $tab_uti = ModelUtilisateur::selectAll();     //appel au modèle pour gerer la BD
+            $controller = static::$object;
+            $view = "list";
+            $pagetitle = "Tout les utilisateurs";
+            $list = File::build_path(array("View","view.php"));
+            require $list;
+        } else {
+            $controller = "utilisateur";
+            $view = "connect";
+            $pagetitle = "Connectez vous pour accéder au panel admin";
+            $list = File::build_path(array("View", "view.php"));
+            require $list;
+        }
     }
 
     public static function read(){
@@ -40,17 +48,26 @@ class ControllerUtilisateur {
     public static function delete(){
         $idUtilisateur = $_GET["idUtilisateur"];
         $tab_uti = ModelUtilisateur::selectAll();
-        if (ModelUtilisateur::delete($idUtilisateur)) {
-            $tab_uti = ModelUtilisateur::selectAll();
-            $controller = self::$object;
-            $view = 'deleted';
-            $pagetitle = 'Utilisateur supprimé';
+        if ($_SESSION['admin'] == 1) {
+            if (ModelUtilisateur::delete($idUtilisateur)) {
+                $tab_uti = ModelUtilisateur::selectAll();
+                $controller = self::$object;
+                $view = 'deleted';
+                $pagetitle = 'Utilisateur supprimé';
+            }
+            else {
+                $controller = self::$object;
+                $view = 'error';
+                $pagetitle = 'Une erreur est survenue lors du delete';
+            }
+        } else {
+            $controller = "utilisateur";
+            $view = "connect";
+            $pagetitle = "Connectez vous pour accéder au panel admin";
+            $list = File::build_path(array("View","view.php"));
+            require $list;
         }
-        else {
-            $controller = self::$object;
-            $view = 'error';
-            $pagetitle = 'Une erreur est survenue';
-        }
+
         require_once File::build_path(array("View", "view.php"));
     }
 
@@ -71,14 +88,23 @@ class ControllerUtilisateur {
             "prenomUtilisateur" => $_POST["prenom"],
             "pseudo" => $_POST["pseudo"],
             "mailUtilisateur" => $_POST["mail"],
-            "motDePasseUtilisateur" => Security::hacher($_POST["motDePasse"])
+            "motDePasseUtilisateur" => Security::hacher($_POST["motDePasse"]),
+            "nonce" => Security::generateRandomHex()
         );
         $pseudo = $_POST["pseudo"];
 
         if ($_POST["motDePasse"] != $_POST["verifMotDePasse"]) {
             $controller = self::$object;
-            $view = 'errorMdp';
+            $view = 'error';
             $pagetitle = 'Erreur mot de passe';
+            echo "Les mots de passes ne correspondent pas";
+            require_once File::build_path(array("View", "view.php"));
+        }
+        else if (filter_var($_POST["mail"], FILTER_VALIDATE_EMAIL) == false){
+            $controller = self::$object;
+            $view = 'error';
+            $pagetitle = 'Erreur mail';
+            echo "L'adresse mail n'est pas valide";
             require_once File::build_path(array("View", "view.php"));
         }
         else if (ModelUtilisateur::save($data)) {
@@ -91,13 +117,37 @@ class ControllerUtilisateur {
             }
             else {
                 $pagetitle= 'Veuillez vous connecter';
+                $view = 'view';
                 require_once File::build_path(array("View", "view.php"));
             }
+
+
+            $mail = 'Veuillez cliquer sur le lien ci-dessous pour vérifier votre compte
+                    https://webinfo.iutmontp.univ-montp2.fr/~gambierc/Projet-Web-S3/index.php?controller=utilisateur&action=validate&nonce=' . $data["nonce"] . '&pseudo=' . $data["pseudo"];
+
+            /*
+            <!DOCTYPE html>
+            <html lang="en" dir="ltr">
+            <head>
+                <meta charset="utf-8">
+            </head>
+
+            <body>
+            <p>Veuillez cliquer sur le lien ci-dessous pour vérifier votre compte</p>
+            <a href="https://webinfo.iutmontp.univ-montp2.fr/~gambierc/Projet-Web-S3/index.php?controller=utilisateur&action=validate&nonce=' . $data["nonce"] . '&pseudo=' . $data["pseudo"] . '"/> <br><br>
+            <p>L\'équipe NFT Factory</p>
+
+            </body>
+            </html>
+             */
+
+            mail($data["mailUtilisateur"], 'Vérification de votre adresse mail', $mail);
+
         }
         else {
             $controller = self::$object;
             $view = 'error';
-            $pagetitle = 'Une erreur est survenue';
+            $pagetitle = 'Une erreur est survenue lors de la création';
             require_once File::build_path(array("View", "view.php"));
         }
     }
@@ -112,11 +162,11 @@ class ControllerUtilisateur {
             $upd = File::build_path(array("View","view.php"));
             require $upd;
         } else {
-            echo "Veuillez vous connecter :";
-            $controller = self::$object;
-            $view = 'connect';
-            $pagetitle = 'Page de connexion';
-            require_once File::build_path(array("View", "utilisateur", "connect.php"));
+            $controller = "utilisateur";
+            $view = "connect";
+            $pagetitle = "Connectez vous pour accéder au panel admin";
+            $list = File::build_path(array("View","view.php"));
+            require $list;
         }
 
     }
@@ -145,7 +195,7 @@ class ControllerUtilisateur {
         else if (!isset($_POST["idUtilisateur"]) || !isset($_POST["nom"]) || !isset($_POST["prenom"]) || !isset($_POST["pseudo"]) || !isset($_POST["mail"]) || !isset($_POST["motDePasse"]) || !ModelUtilisateur::update($data)) {
             $controller = self::$object;
             $view = 'error';
-            $pagetitle = 'Une erreur est survenue';
+            $pagetitle = 'Une erreur est survenue lors de l\'update';
             require_once File::build_path(array("View", "view.php"));
         }
 
@@ -165,37 +215,45 @@ class ControllerUtilisateur {
     }
 
     public static function connected(){
-            if (!isset($_GET['pseudo']) || !isset($_GET['motDePasse'])) {
+            if (!isset($_POST['pseudo']) || !isset($_POST['motDePasse'])) {
                 $controller = self::$object;
                 $view = 'error';
                 $pagetitle = 'Erreur de connexion';
                 require_once File::build_path(array("View", "view.php"));
             }
 
-            $pseudo = $_GET['pseudo'];
-            $mdp = Security::hacher($_GET['motDePasse']);
+            $pseudo = $_POST['pseudo'];
+            $mdp = Security::hacher($_POST['motDePasse']);
 
             if(ModelUtilisateur::checkPassword($pseudo, $mdp)) {
-                $_SESSION['login'] = $pseudo;
-                $user = ModelUtilisateur::selectWithPseudo($pseudo);
-                $_SESSION['idUser'] = $user->get('idUtilisateur');
+                if (ModelUtilisateur::checkMail($pseudo)) {
+                    $_SESSION['login'] = $pseudo;
 
-                if (ModelUtilisateur::isAdmin($pseudo)) {
-                    $_SESSION['admin'] = 1;
-                    $controller = 'Admin';
-                    $view = 'list';
-                    $pagetitle = 'Bienvenue ' . $pseudo . ' !';
-                    $action='afficher';
-                    require_once File::build_path(array("View", "view.php"));
-                }
-                else {
-                    $_SESSION['admin'] = 0;
+                    if (ModelUtilisateur::isAdmin($pseudo)) {
+                        $_SESSION['admin'] = 1;
+                        $controller = 'admin';
+                        $view = 'list';
+                        $pagetitle = 'Bienvenue ' . $pseudo . ' !';
+                        $action='afficher';
+                        require_once File::build_path(array("View", "view.php"));
+                    }
+                    else {
+                        $_SESSION['admin'] = 0;
+                        $controller = 'static';
+                        $view = 'home';
+                        $pagetitle = 'Bienvenue ' . $pseudo . ' !';
+                        require_once File::build_path(array("View", "view.php"));
+                    }
+                } else {
+
                     $controller = self::$object;
-                    $view = 'marketPlace';
-                    $pagetitle = 'Bienvenue ' . $pseudo . ' !';
-                    $tab_prod = ModelProduit::selectAll();
+                    $view = 'error';
+                    $pagetitle = 'Mauvais mot de passe';
+                    echo "Veuillez vérifier votre adresse mail";
                     require_once File::build_path(array("View", "view.php"));
                 }
+                
+                
             }
             else {
                 $controller = self::$object;
@@ -205,10 +263,37 @@ class ControllerUtilisateur {
             }
     }
 
+    public static function validate() {
+        if (!isset($_GET['pseudo']) || !isset($_GET['nonce'])) {
+            $controller = self::$object;
+            $view = 'error';
+            $pagetitle = 'Erreur de vérification du mail';
+            require_once File::build_path(array("View", "view.php"));
+        }
+
+        $requete = Model::getPDO()->query('SELECT nonce FROM Utilisateur WHERE pseudo="' . $_GET['pseudo'] . '"'); //gérer le cas ou le pseudo est pas bon et ou le nonce est déjà à null
+        if ($requete->fetchColumn() == $_GET['nonce']) {
+            $change = Model::getPDO()->prepare('UPDATE Utilisateur SET nonce=NULL WHERE pseudo="' . $_GET['pseudo'] . '"');
+            $change->execute();
+            $controller = self::$object;
+            $view = 'marketPlace';
+            $pagetitle = 'Merci d\'avoir vérifié votre mail!';
+            $tab_prod = ModelProduit::selectAll();
+            require_once File::build_path(array("View", "view.php"));
+        } else {
+            $controller = self::$object;
+            $view = 'error';
+            $pagetitle = 'Erreur de vérification de l\'adresse mail';
+            require_once File::build_path(array("View", "view.php"));
+        }
+        
+    }
+
     public static function marketPlace() {
         $controller = static::$object;
         $view = "marketPlace";
         $pagetitle = "Market place";
+        $tab_cat = ModelCategorie::selectAll();
 
         $tab_prod = ModelProduit::selectAll();
 
@@ -216,9 +301,18 @@ class ControllerUtilisateur {
     }
 
     public static function afficherPanier() {
-        $controller = static::$object;
-        $view = "panier";
-        $pagetitle = "Panier";
+        if ($_SESSION['login'] != NULL) {
+            $controller = static::$object;
+            $view = "panier";
+            $pagetitle = "Panier";
+        } else {
+            $controller = "utilisateur";
+            $view = "connect";
+            $pagetitle = "Connectez vous pour accéder au panel admin";
+            $list = File::build_path(array("View","view.php"));
+            require $list;
+        }
+
 
         require_once File::build_path(array("View","view.php"));
     }
@@ -226,6 +320,7 @@ class ControllerUtilisateur {
 
 
     public static function ajouterAuPanier(){
+        $tab_cat = ModelCategorie::selectAll();
         if(empty($_SESSION['panier'])){
             $_SESSION['panier'] = array();
         }
@@ -271,8 +366,9 @@ class ControllerUtilisateur {
     public static function deconnect() {
         session_unset();
         $controller = self::$object;
-        $view = 'detail';
+        $view = 'marketPlace';
         $pagetitle = 'Déconnexion réussie';
+        $tab_prod = ModelProduit::selectAll();
         require_once File::build_path(array("View", "view.php"));
     }
 }
